+++ libs/storybook/.storybook/preview.js
@@ -12,6 +12,20 @@ setCompodocJson(docJson);
 export const parameters = {
   html: {
     root: '[ng-version]',
+    prettier: {
+      htmlWhitespaceSensitivity: 'ignore',
+      bracketSameLine: true,
+      singleAttributePerLine: false,
+      printWidth: 140,
+    },
+    removeComments: true,
+    transform: (code) => {
+      // Remove attributes `_nghost` and `ng-reflect` injected by Angular:
+      return code.replace(
+        /(?:_nghost|ng-reflect|_ngcontent|data-cy|dataattribute|cving|data-chromatic).*?="[\S\s]*?"/g,
+        ''
+      );
+    },
   },
   docs: {
     extractComponentDescription: (_, { notes }) => {
@@ -20,6 +34,26 @@ export const parameters = {
       }
       return null;
     },
+    source: {
+      state: 'open',
+    },
+  },
+  backgrounds: {
+    default: 'Default',
+    values: [
+      {
+        name: 'Default',
+        value: '#ffffff',
+      },
+      {
+        name: 'Dark',
+        value: '#13100d',
+      },
+      {
+        name: 'Gray',
+        value: '#f0f0f2',
+      },
+    ],
   },
   options: {
     // stories inside one file will be sorted in alphabetical order because of this bug https://github.com/storybookjs/storybook/issues/18659