+++ libs/ui/src/lib/button/button.component.stories.ts
@@ -1,6 +1,7 @@
 import { Story, Meta } from '@storybook/angular';
 import notes from './button.component.md';
 import { ButtonComponent } from './button.component';
+import { storybookIconsNames } from '../icons/storybook-icons';
 
 export default {
   title: 'Angular/Button',
@@ -11,6 +12,11 @@ export default {
       name: 'Appearance',
       options: ['primary', 'secondary', 'text'],
       control: { type: 'inline-radio' },
+    },
+    iconName: {
+      name: 'Icon name',
+      options: storybookIconsNames,
+      control: { type: 'select' },
     },
     content: {
       name: 'Content',
@@ -31,7 +37,7 @@ const Template: Story<ButtonComponent> = (args: ButtonComponent) => ({
   props: args,
   /* template */
   template: `
-    <cvi-ng-button [disabled]="disabled" [size]="size" [appearance]="appearance">{{ content }}</cvi-ng-button>
+    <cvi-ng-button [disabled]="disabled" [size]="size" [appearance]="appearance" [iconName]="iconName" [iconPosition]="iconPosition"[iconHeight]="iconHeight">{{ content }}</cvi-ng-button>
   `,
 });
 
@@ -53,18 +59,41 @@ Text.args = {
   appearance: 'text',
 };
 
-const TemplateTextButtonWithIcon: Story<ButtonComponent> = (
+export const WithIcon = Template.bind({});
+WithIcon.args = {
+  iconName: 'add',
+  iconPosition: 'after',
+};
+
+export const WithIconSmall = Template.bind({});
+WithIconSmall.args = {
+  size: 's',
+  iconName: 'add',
+  iconPosition: 'after',
+};
+
+export const WithIconBefore = Template.bind({});
+WithIconBefore.args = {
+  iconName: 'add',
+  iconPosition: 'before',
+};
+
+const TemplateWithCustomColor: Story<ButtonComponent> = (
   args: ButtonComponent
 ) => ({
   props: args,
+  styles: [
+    `:host {
+      --cvi-button-color: var(--cvi-color-jasper-10);
+      --cvi-button-color--hover: var(--cvi-color-jasper-12);
+    }`,
+  ],
   /* template */
   template: `
-    <cvi-ng-button appearance="text">
-      <cvi-ng-labeled-icon name="add" iconPosition="before" verticalAlignment="center" [iconHeight]="16" [gap]="2">
-        {{ content }}
-      </cvi-ng-labeled-icon>
+    <cvi-ng-button [disabled]="disabled" [size]="size" [appearance]="appearance" [iconName]="iconName" [iconPosition]="iconPosition" [iconHeight]="iconHeight">
+      {{ content }}
     </cvi-ng-button>
   `,
 });
 
-export const TextButtonWithIcon = TemplateTextButtonWithIcon.bind({});
+export const WithCustomColor = TemplateWithCustomColor.bind({});