+++ libs/ui/src/lib/input/input.component.stories.ts
@@ -19,6 +19,11 @@ export default {
       options: storybookIconsNames,
       control: { type: 'select' },
     },
+    validationType: {
+      name: 'ValidationType',
+      options: [null, 'positiveNumbers'],
+      control: { type: 'select' },
+    },
   },
   args: {
     placeholder: 'Username',
@@ -35,7 +40,8 @@ const Template: Story<InputComponent> = (args: InputComponent) => ({
       <cvi-ng-input [placeholder]="placeholder"
                     [disabled]="disabled"
                     [suffixIconName]="suffixIconName"
-                    [htmlId]="htmlId"></cvi-ng-input>
+                    [htmlId]="htmlId"
+                    [validationType]="validationType"></cvi-ng-input>
     </cvi-ng-form-item>
   `,
 });
@@ -60,12 +66,22 @@ const FormTemplate: Story<InputComponent> = (args: InputComponent) => {
     console.log(formValue);
   }
 
+  function disableInput() {
+    form.controls.item.disable();
+  }
+
+  function enableInput() {
+    form.controls.item.enable();
+  }
+
   return {
     props: {
       ...args,
       form,
       selectedValue,
       onSubmit,
+      disableInput,
+      enableInput,
     },
     /* template */
     template: `
@@ -74,9 +90,15 @@ const FormTemplate: Story<InputComponent> = (args: InputComponent) => {
                           [htmlId]="htmlId">
           <cvi-ng-input formControlName="item"
                         [placeholder]="placeholder"
-                        [disabled]="disabled"
-                        [htmlId]="htmlId"></cvi-ng-input>
+                        [htmlId]="htmlId"
+                        [validationType]="validationType"></cvi-ng-input>
         </cvi-ng-form-item>
+        <cvi-ng-track layout="flex" horizontalAlignment="justify" gap="3">
+            <cvi-ng-button data-cy="disable-button" (click)="disableInput()">Disable input</cvi-ng-button>
+            <cvi-ng-button data-cy="enable-button" (click)="enableInput()">Enable input</cvi-ng-button>
+        </cvi-ng-track>
+        <div></div>
+
       </form>
       <div>Inserted value: {{selectedValue()}}</div>
     `,
@@ -96,7 +118,8 @@ const CharacterCounterTemplate: Story<InputComponent> = (
                     [disabled]="disabled"
                     [htmlId]="htmlId"
                     cviNgCharacterCounter
-                    [maxChars]="10"></cvi-ng-input>
+                    [maxChars]="10"
+                    [validationType]="validationType"></cvi-ng-input>
     </cvi-ng-form-item>
   `,
 });